+++ home-assistant-plugin/custom_components/xknx/const.py
@@ -35,7 +35,7 @@ class ColorTempModes(Enum):
 
     ABSOLUTE = "DPT-7.600"
     RELATIVE = "DPT-5.001"
-    RELATIVE_TEMP = "DPT-9.002"
+    ABSOLUTE_TEMP = "DPT-9.002"
 
 
 class SupportedPlatforms(Enum):